Has_rich_text for description
