+++ app/controllers/enrollments_controller.rb
@@ -10,7 +10,7 @@ class EnrollmentsController < ApplicationController
       flash[:alert] = I18n.t('course_is_already_enrolled')
     else
       @course.enrollments.create(student: current_student)
-      Mailers::EnrollmentMailer.new_students_enrollment(@course.instructor, @course, current_student).deliver_now
+      Instructors::InstructorsMailer.new_students_enrollment(@course.instructor, @course, current_student).deliver_now
       flash[:notice] = I18n.t('course_is_successfully_enrolled')
     end
     redirect_to @course
