+++ app/jobs/course/new_course_email_job.rb
@@ -4,6 +4,6 @@ class Course::NewCourseEmailJob
   def perform(instructor_id, course_id)
     instructor = Instructor.find(instructor_id)
     course = Course.find(course_id)
-    Mailers::CourseMailer.new_course_created(instructor, course).deliver_now
+    Instructors::InstructorsMailer.new_course_created(instructor, course).deliver_now
   end
 end
