--- app/mailers/mailers/course_mailer.rb
@@ -1,11 +1,0 @@
-module Mailers
-  class CourseMailer < ApplicationMailer
-    def new_course_created(instructor, course)
-      @course = course
-      mail(
-        to: instructor.email,
-        subject: course.name
-      )
-    end
-  end
-end
