+++ app/mailers/students/students_mailer.rb
@@ -1,5 +1,5 @@
-module Mailers
-  class StartCourseMailer < ApplicationMailer
+module Students
+  class StudentsMailer < ApplicationMailer
     def course_start(course, student)
       @course = course
       @student = student
