+++ app/services/course_starting_service.rb
@@ -14,7 +14,7 @@ class CourseStartingService
 
   def send_notification_to_students
     @course.students.each do |student|
-      Mailers::StartCourseMailer.course_start(@course, student).deliver_now
+      Students::StudentsMailer.course_start(@course, student).deliver_now
     end
   end
 end
