+++ app/views/courses/edit.html.erb
@@ -10,7 +10,7 @@
                           class: 'form-control' %>
     </div>
     <div data-mdb-input-init class="form-outline mb-2">
-      <%= form.text_field :description,
+      <%= form.rich_text_area :description,
                           autofocus: true,
                           autocomplete: 'description',
                           placeholder: 'course description',
