+++ app/views/courses/show.html.erb
@@ -1,6 +1,6 @@
 <h3><%= @course.name %></h3>
 <div class="description">
-  " <%= @course.description %>"
+  <%= @course.description %>
 </div>
 <div>
   <strong>Status:</strong>
@@ -20,7 +20,7 @@
   <% end %>
 </div>
 
-<% if  policy(Course).student_list? %>
+<% if  policy(Course).student_list? && @course.has_students?%>
   <div class="table-responsive mt-5">
     <table class="table table-striped table-hover">
       <thead>
