+++ app/views/instructors/instructors_mailer/new_course_created.html.erb
