+++ app/views/instructors/instructors_mailer/new_students_enrollment.html.erb
