+++ app/views/students/students_mailer/course_start.html.erb
