+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2025_01_18_115953) do
+ActiveRecord::Schema.define(version: 2024_12_21_161333) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -51,17 +51,6 @@ ActiveRecord::Schema.define(version: 2025_01_18_115953) do
     t.bigint "blob_id", null: false
     t.string "variation_digest", null: false
     t.index ["blob_id", "variation_digest"], name: "index_active_storage_variant_records_uniqueness", unique: true
-  end
-
-  create_table "comments", force: :cascade do |t|
-    t.bigint "courses_id", null: false
-    t.bigint "instructors_id", null: false
-    t.bigint "students_id", null: false
-    t.datetime "created_at", precision: 6, null: false
-    t.datetime "updated_at", precision: 6, null: false
-    t.index ["courses_id"], name: "index_comments_on_courses_id"
-    t.index ["instructors_id"], name: "index_comments_on_instructors_id"
-    t.index ["students_id"], name: "index_comments_on_students_id"
   end
 
   create_table "courses", force: :cascade do |t|
@@ -133,9 +122,6 @@ ActiveRecord::Schema.define(version: 2025_01_18_115953) do
 
   add_foreign_key "active_storage_attachments", "active_storage_blobs", column: "blob_id"
   add_foreign_key "active_storage_variant_records", "active_storage_blobs", column: "blob_id"
-  add_foreign_key "comments", "courses", column: "courses_id"
-  add_foreign_key "comments", "instructors", column: "instructors_id"
-  add_foreign_key "comments", "students", column: "students_id"
   add_foreign_key "courses", "instructors"
   add_foreign_key "enrollments", "courses"
   add_foreign_key "enrollments", "students"
