+++ package.json
@@ -4,10 +4,12 @@
   "dependencies": {
     "@popperjs/core": "^2.11.8",
     "@rails/actioncable": "^6.0.0",
+    "@rails/actiontext": "^6.1.710",
     "@rails/activestorage": "^6.0.0",
     "@rails/ujs": "^6.0.0",
     "@rails/webpacker": "5.4.4",
     "bootstrap": "^5.3.3",
+    "trix": "^1.2.0",
     "turbolinks": "^5.2.0",
     "webpack": "^4.46.0",
     "webpack-cli": "^3.3.12"
